tcp_proxy: fixes a cx leak in the TCP Proxy when receive_before_connect is enabled (#42024)
This PR fixes a connection leak in the TCP proxy when the
`receive_before_connect` feature is enabled and the downstream
connection closes before the upstream connection is established.

After this, the TCP Proxy should properly propagates the end-of-stream
signal to the upstream connection even when no data gets received from
the downstream, preventing upstream connection leaks.

Fix https://github.com/envoyproxy/envoy/issues/42006

---

**Commit Message**: tcp_proxy: fixes a cx leak in the TCP Proxy when
receive_before_connect is enabled
**Additional Description:** Fixed a connection leak in the TCP proxy
when the `receive_before_connect` feature is enabled and the downstream
connection closes before the upstream connection is established.
**Risk Level:** Low
**Testing:** Added Tests
**Docs Changes:** N/A
**Release Notes:** Added

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>