network: Defer getting current time until after receiving packets
QUICHE added a DCHECK[1] to one of QuicBandwidth's multiplication
operators, which caused one of Envoy's QUIC integration test to flakily
fail[2]. When the test fails, QUICHE's BandwidthSampler sees ACK
timestamps that precede the corresponding packet's send time!

We think the root cause lies in Utility::readPacketsFromSocket(). It's
plausible that a sending thread preempts a receiver thread between calls
to monotonicTime() and readFromSocket(). The sequence of events would
look like this:

        receiver thread: ack_time = monotonicTime()
          sender thread: send_time = monotonicTime()
        receiver thread: read packet(s)
          sender thread: send packets

This commit reverses the receiver's order of operations. The receiver
now reads packets *before* getting the current time. This way, when the
receiver is preempted between operations, the only effect is that
ack_time may be slightly larger. It's acceptable to have positive error
on ack_time because it preserves the `send_time <= ack_time` property.

[1]: https://github.com/google/quiche/commit/0e29fdb0e661ca7c978e550ece6e38d3dc342e47
[2]: QuicHttpIntegrationTests/QuicHttpIntegrationTest.MultipleQuicConnectionsNoBPF/IPv6

Signed-off-by: Dan McArdle <dmcardle@google.com>