ext_authz: fixed a bug to honor filter_enabled_metadata when the filter is marked with disabled=true (#41937)
## Description

This PR fixes a bug to honor ``filter_enabled_metadata`` when the filter
is marked with ``disabled: true``. The filter is now instantiated when
``filter_enabled_metadata`` is configured, allowing it to check dynamic
metadata at runtime and enable itself conditionally.

Fix https://github.com/envoyproxy/envoy/issues/41501

---

**Commit Message:** ext_authz: fixed a bug to honor
filter_enabled_metadata when the filter is marked with disabled=true
**Additional Description:** Fixes a bug to honor
``filter_enabled_metadata`` when the filter is marked with ``disabled:
true``.
**Risk Level:** Low
**Testing:** Added Tests
**Docs Changes:** N/A
**Release Notes:** Added

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>