reverse_tunnel:  prevent TLS close_notify alert when duplicating reverse connection sockets (#42110)
## Description

Commit Message: prevent TLS close_notify alert when duplicating reverse
connection sockets

**Additional Description:**
When a reverse connection is established, the socket is
[duplicated](https://github.com/envoyproxy/envoy/blob/af705b7dae2c0c589cde096e335f17cac7ad33b6/source/extensions/bootstrap/reverse_tunnel/downstream_socket_interface/reverse_connection_io_handle.cc#L259)
and passed to the reverse
connection listener, and the original socket is closed. When TLS is used
for the upstream cluster, a TLS close_notify alert message is sent to
upstream envoy after socket duplication and closure, which upstream
envoy replies to. This reply is read on the duplicated socket FD. The
DownstreamReverseConnectionIOHandle [reads this TLS Alert
](https://github.com/envoyproxy/envoy/blob/af705b7dae2c0c589cde096e335f17cac7ad33b6/source/extensions/bootstrap/reverse_tunnel/downstream_socket_interface/downstream_reverse_connection_io_handle.cc#L35)instead
of
RPINGs, therefore breaking the health check mechanism.

**Solution:**
Set SSL quiet shutdown mode on the connection before closing it, this
prevents the close_notify alert from being sent when the original
connection is closed, while the duplicated socket remains operational
for both keepalives and application data
transfer.

Fix https://github.com/envoyproxy/envoy/issues/42116

---

**Risk Level:** Low
**Testing:** Added Unit Tests
**Docs Changes:** N/A
**Release Notes:** N/A
**Platform Specific Features:** N/A

---------

Signed-off-by: Basundhara Chakrabarty <basundhara.c@nutanix.com>