docs: enhance the docs around the retry_policy for gRPC (#42115)
## Description

Currently, the doc for [retry_policy](https://arc.net/l/quote/rjzyalhz)
for gRPC is a bit confusing. We have multiple level of retries from
`envoy_grpc`, `google_grpc` and an outer `retry_policy`. This is an
attempt to add some more clarification on the Retry Policy and which one
is applicable for ExtAuthZ.

---

**Commit Message:** docs: enhance the docs around the retry_policy for
gRPC
**Additional Description:**
**Risk Level:** N/A
**Testing:** CI
**Docs Changes:** N/A
**Release Notes:** N/A

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>