ext_proc: fix a call stats logging issue (#42055)
In STREAMED or FULL_DUPLEX_STREAMED mode, the body or trailers stats are
sometime not logged.
This is due to a bug in ext_proc logging implementation which only do
the logging when the call_start_time_ is set. However, in STREAMED or
FULL_DUPLEX_STREAMED body mode cases, there could be multiple data
chunks are sent to the server, and the 1st response will clear the
call_start_time_. This causes the rest body responses won't be logged.
The solution is to not clear call_start_time_ when a response is
received.

---------

Signed-off-by: Yanjun Xiang <yanjunxiang@google.com>