upstream_filter: fix an issue with ext_proc filter configured in the upstream filter chain (#41879)
Fixes an issue when an HTTP filter is in the upstream filter chain, and
its encodeHeaders() method is returning StopIteration() , and route
retry policy is configured with retry on 5xx. In this case, when a 503
response is received, onUpstreamHeaders() will be called first, and the
UpstreamRequest object will be removed from the upstream_request_ list.
onUpstreamData() will be called right after, and should just return.

---------

Signed-off-by: Yanjun Xiang <yanjunxiang@google.com>