transport_socket(http_11_proxy): Flush write buffer after CONNECT response
Fixes buffering problem in scenario where bytes written to a connection
after the initial HTTP CONNECT request is sent, but before the response
is received from the upstream proxy. In that case, writes are paused
until the CONNECT response is received, but never unpaused which can
cause buffered writes to be stuck indefinitely.

This patch flushes the write buffer after a successful CONNECT response
is received.

Signed-off-by: Tony Allen <txallen@google.com>