tcp_proxy: add upstream_connect_trigger to control connection establishment (#41696)
## Description

This PR adds a new `upstream_connect_trigger` in TCP Proxy to control
how TCP Proxy establishes connections with upstream. In some use-cases
we want to delay the connection until we receive the data or let it
complete the TLS handshake which is not possible today.

---

**Commit Message:** tcp_proxy: add upstream_connect_trigger to control
connection establishment
**Additional Description:** Adds a new `upstream_connect_trigger` in TCP
Proxy to control how TCP Proxy establishes connections with upstream.
**Risk Level:** Low
**Testing:** Added Unit + Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>