router: added retry_policy to HttpProtocolOptions to support retries (#42077)
## Description

This PR adds a new `retry_policy` to `HttpProtocolOptions` to support
retries on the cluster level so that all the routes that goes to this
upstream cluster gets it.

---

**Commit Message:** router: added retry_policy to HttpProtocolOptions to
support retries
**Additional Description:** Adds a new `retry_policy` to
`HttpProtocolOptions` to support retries on the cluster level so that
all the routes that goes to this upstream cluster gets it.
**Risk Level:** Low
**Testing:** Added Unit & Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>