Add configuration for MCP router filter (#42275)
Extension for abstracting multiple MPC servers as one. Capabilities,
tools and resources of remote MCP servers are combined and presented as
a single MCP server to the client. This allows a unified policy to be
applied to multiple remote servers.

Part of #39174

Risk Level: none
Testing: n/a
Docs Changes: yes
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Yan Avlasov <yavlasov@google.com>