mobile: Use a Mock DNS resolver in integration tests
This change does two things:
  1. Creates a new MockDnsResolver that can be used for mobile
     integration tests setting the host to a non-local host.
  2. Updates ClientIntegrationTest to use the MockDnsResolver for tests
     that attempt to connect to a non-existant host.

The MockDnsResolver treats localhost addresses like a standard DNS
resolver would, but for the configured "non-existant" hosts, it returns
an empty DNS record (EAI_NONAME) without attempting to go on the wire to
fetch the host's DNS records. This saves test time and makes for a more
hermetic test.

Signed-off-by: Ali Beyad <abeyad@google.com>