contrib: add a new postgres inspector filter (#40803)
## Description

We have a few use-cases where we'd like to inspect the SNI of the
incoming Postgres connections and then based on the SNI go connect to
the Database which is not possible today.

This PR adds a new Postgres Inspector filter in Contrib which could be
used to populate the protocol as "postgres" and be chained with TLS
Inspector to extract the SNI. SNI could later be used in TCP Proxy to
connect to a unique Database.

---

**Commit Message:** contrib: add a new postgres inspector filter
**Additional Description:** Added a new Contrib filter for Postgres
Inspector which would be responsible for inspecting the incoming TLS PG
connections.
**Risk Level:** Low
**Testing:** Added Unit & Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>