Include openssh-client in devcontainer (#42437)
`git push origin $BRANCHNAME` doesn't work from
inside a devcontainer without this change, due to the absence of `ssh`.
An alternative way to achieve the same end would be to include the ssh
client in the base image.

Testing: Rebuilt a devcontainer after this change and it went from not

Signed-off-by: Raven Black <ravenblack@dropbox.com>