docs: add a note on Network ExtAuthZ usage with TCP Proxy to delay connect (#42393)
## Description

This PR add documentation on how `upstream_connect_mode` in the TCP
Proxy filter could be used to make sure that we wait for ExtAuthZ to
complete before opening a connection to the backend.

---

**Commit Message:** docs: add a note on Network ExtAuthZ usage with TCP
Proxy to delay connect
**Additional Description:** Added documentation on how
`upstream_connect_mode` in the TCP Proxy filter could be used to make
sure that we wait for ExtAuthZ to complete before opening a connection
to the backend.
**Risk Level:** N/A
**Testing:** CI
**Docs Changes:** Added
**Release Notes:** N/A

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>