docs: add a note on Network RBAC usage with TCP Proxy to delay connect (#42392)
## Description

This PR add documentation on how `upstream_connect_mode` in the TCP
Proxy filter could be used to make sure that we wait for RBAC
enforcement to complete before opening a connection to the backend.

Fix https://github.com/envoyproxy/envoy/issues/9023

---

**Commit Message:** docs: add a note on Network RBAC usage with TCP
Proxy to delay connect
**Additional Description:** Added documentation on how
`upstream_connect_mode` in the TCP Proxy filter could be used to make
sure that we wait for RBAC enforcement to complete before opening a
connection to the backend.
**Risk Level:** N/A
**Testing:** CI
**Docs Changes:** Added
**Release Notes:** N/A

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>