ext_authz: add error_response for gRPC ExtAuthZ to return custom error response (#42099)
## Description

Similar to
[denied_response](https://www.envoyproxy.io/docs/envoy/latest/api-v3/service/auth/v3/external_auth.proto#service-auth-v3-checkresponse),
this PR adds a new `error_response` on the gRPC ExtAuthZ service to
provide a way to send back an `error_response` with a custom status
code, headers and/or body from gRPC ExtAuthZ servers back to Envoy.

It's useful in cases where we are not explicitly "denying" the request
but just communicating that the server could be overloaded or the
processing has failed for some reason and suggesting the user to retry
after some time if needed.

Fix https://github.com/envoyproxy/envoy/issues/42098

---

**Commit Message:** ext_authz: add error_response for gRPC ExtAuthZ to
return custom error response
**Additional Description:** Adds a new `error_response` on the gRPC
ExtAuthZ service to provide a way to send back an `error_response` with
a custom status code, headers and/or body from gRPC ExtAuthZ servers
back to Envoy.
**Risk Level:** Low
**Testing:** Added Unit + Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>