Add requested application protocols to ConnectionInfoProvider (#42339)
This change introduces methods to get and set the requested application
protocols (ALPNs) on the ConnectionInfoProvider interface. The storage
for these protocols is moved from ConnectionSocketImpl to
ConnectionInfoProviderImpl. Handling of ALPNs now parallels that for SNI
and ensures that both are accessible from StreamInfo.

Also adding a MockConnectionInfoProvider class for testing since I
noticed that one does not exist.

Risk Level: low
Testing: unit tests

---------

Signed-off-by: Elisha Ziskind <eziskind@google.com>