geoip: allow extracting the client IP address from a custom request header (#42416)
## Description

This PR adds a new `ip_address_header` which could be used to provide
the client IP address instead of computing it using XFF hops. It's
valuable in cases where we trust the first layer to correctly determine
the client IP and then provide that as a header.

---

**Commit Message:** geoip: allow extracting the client IP address from a
custom request header
**Additional Description:** Adds a new `ip_address_header` which could
be used to provide the client IP address instead of computing it using
XFF hops.
**Risk Level:** Low
**Testing:** Added Unit + Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>