Fix the stale node cleanup causing segfault upstream socket manager (#42635)
## Commit Message
Fix the stale node cleanup causing segfault upstream socket manager

## Additional Description
If the last socket for a node has some error, we cleanup the stale node
from the accepted_reverse_connections_ map which causes a segfault cuz
the for loop holds an invalid iterator.

Pre increment of iterator solves this. The test passes now and causes a
crash for the previous code.

Signed-off-by: aakugan <aakashganapathy2@gmail.com>