http: apply the buffer limit when route is refrshed (#42607)
In the previous implementation, the `request_body_buffer_limit` will be
applied when the `decodeHeaders` of router filter is called. That's say
if one filter return stop when processing the headers, the buffer limit
of route will not works.

This PR will apply the limit when the route is matched.

Risk Level: low.
Testing: unit.
Docs Changes: n/a.
Release Notes: added.
Platform Specific Features: n/a.

---------

Signed-off-by: wbpcode/wangbaiping <wbphub@gmail.com>