Avoid multimap::find unspecified behavior
When a multimap has multiple entries with the same key, calling multimap::find(key) returns an unspecified element. 

Historically, this returns the first matching element. However, this is not guaranteed, and recent libc++ changes make this return an arbitrary element.

Using equal_range() is a replacement that will preserve the current behavior. The behavior of this is guaranteed to return the first match instead of an arbitrary one.

Signed-off-by: Jordan Rupprecht <rupprecht@google.com>