reverse_tunnels: Add in support for the reverse tunnel reporting by adding hook points and interface for the reporter (#42647)
## Commit Message 
Add in support for the reverse tunnel reporting by adding hook points
and interface for the reporter

## Additional Description
Adds an interface for reporters to extend to report the reverse tunnels.
Also added in the hooks needed on the extension, socket manager and the
filter to report the tunnels with the context to report to the reporter
(configured using yaml).

We are reporting the tenant_id on connections because we have plans to
possibly use them for validation later on in the request lifecycle. This
data isnt available at the upstream_socket_manager hence we dont report
the tenant_id on disconnection

Also added tests for the exercised new code paths.

Signed-off-by: aakugan <aakashganapathy2@gmail.com>