access_log: add a new COALESCE substitution formatter (#42711)
## Description

This PR adds a new `COALESCE` substitution formatter that evaluates
multiple formatter operators in sequence and returns the first non-null
result. This could be used to have a fallback behavior such as using SNI
when available but falling back to the `:authority` header when SNI is
not available.

**Example:**

```
%COALESCE(
  {
    "operators": [
      "REQUESTED_SERVER_NAME",
      {"command": "REQ", "param": ":authority"},
      {"command": "REQ", "param": "x-envoy-original-host"}
    ]
  }
)%
```

---

**Commit Message:** access_log: add a new **COALESCE** substitution
formatter
**Additional Description:** Adds a new `COALESCE` substitution formatter
that evaluates multiple formatter operators in sequence and returns the
first non-null result.
**Risk Level:** Low
**Testing:** Added Unit + Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>