Rename Envoy::Matcher::MatchResult to ActionMatchResult
This is in preparation to promote the Envoy::Matcher::FieldMatchResult (which is more generic in that it does not imply an action if matched) to replace the regular MatchResult.

This is the first step of changing InputMatcher to return the MatchResult (née) FieldMatchResult type, permitting it to return the MatchResult::insufficientData rather than just a boolean.

Signed-off-by: Matt Leon <mattleon@google.com>