dynamic_modules: add support for disabling HTTP per-route filter (#42766)
## Description

In this PR we are adding a new `disabled` flag to the per-route
configuration for the dynamic modules HTTP filter which could be used to
disable the dynamic modules on a per-route basis if needed. This matches
the functionality we offer in other filters like ExtAuthZ, etc.

---

**Commit Message:** dynamic_modules: add support for disabling HTTP
per-route filter
**Additional Description:** Added a new `disabled` flag to the per-route
configuration for the dynamic modules HTTP filter which could be used to
disable the dynamic modules on a per-route basis
**Risk Level:** Low
**Testing:** Added Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>