Avoid multimap::find unspecified behavior (#42705)
Commit Message: When a multimap has multiple entries with the same key,
calling multimap::find(key) returns an unspecified element.

Historically, this returns the first matching element. However, this is
not guaranteed, and recent libc++ changes make this return an arbitrary
element.

Using equal_range() is a replacement that will preserve the current
behavior. The behavior of this is guaranteed to return the first match
instead of an arbitrary one.

Additional Description:
Risk Level: Low
Testing: Unit tests
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Jordan Rupprecht <rupprecht@google.com>